+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - **Made _Tag Finder_/_Highlight P.O.I.'s_ cross marks have a fixed scale**
 - **Added support for an optional `DSMILEST` sound for _Milestone-Completion Announcements_**
 - **Made the weapon carousel partially customizable through NUGHUD**
+- **Improved palettization of true-color PNG graphics**
 - **Made initialization of _Radial Fog_ not trigger dynamic resolution changes**
 - **Changed the default values of the following settings**: [^2]
   - Set _Auto Save Interval_ to 300s, enabling periodic auto-saves